+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add cloud-init config file to install and configure apache with and without https
 - add *default_authorized_keys* file and corresponding folder readme
 - add *terraform.tfvars* template
+- add subnets resource config *subnets.tf*
+- add security resource config *security.tf*
 
 ### Changed